some cleanup

+++ sw/simulation/simulation_thread.h
@@ -57,13 +57,14 @@ vector<float> generate_random_vector_zeromean(const int &length)
 {
   // Generate the random vector
   vector<float> v(length, 0);
-  for (int i = 0; i < length; i++) {
+  for (uint8_t i = 0; i < length; i++) {
     v[i] = getrand_float(-0.5, 0.5);
   }
 
   // Adjust to zero mean
   vector<float> temp = v;
-  for (int i = 0; i < length; i++) {
+  for (uint8_t i = 0; i < length; i++)
+  {
     v[i] = v[i] - vector_mean(temp);
   }
